+++ packages/playwright-ariada/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @ariada-org/playwright-ariada
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [258c8da]
+- Updated dependencies [d7a10f5]
+- Updated dependencies [3cc2ae5]
+  - @ariada-org/cli@0.3.1
+  - @ariada-org/core-playwright@0.4.0
+  - @ariada-org/core-engine@0.3.1
+
 ## 0.1.1
 
 ### Patch Changes